Split constants and changed visibility
