+++ crates/ekore/src/math_constants.rs
@@ -1,0 +1,17 @@
+//! Math Constants.
+
+/// Riemann zeta function at z = 2.
+///
+/// $\zeta(2) = \pi^2 / 6$.
+pub(crate) const ZETA2: f64 = 1.6449340668482264;
+
+/// Riemann zeta function at z = 3.
+pub(crate) const ZETA3: f64 = 1.2020569031595942;
+
+/// Riemann zeta function at z = 4.
+///
+/// $\zeta(4) = \pi^4 / 90$.
+pub(crate) const ZETA4: f64 = 1.082323233711138;
+
+/// Riemann zeta function at z = 5.
+pub(crate) const ZETA5: f64 = 1.03692775514337;
